feat: add `docker-compose.dev.yaml` for development environment
This commit adds the `docker-compose.dev.yaml` file, which defines the services and configurations for the development environment. It includes services for MongoDB, Redis, and RabbitMQ, along with their respective configurations and volumes. This file will be used to set up the development environment with the necessary dependencies and network configurations.

[skip deploy] [skip sdk]